+++ docs/whatsnew/changelog.rst
@@ -20,12 +20,12 @@ New features
 ~~~~~~~~~~~~
 .. Add here new public features (do not delete this comment)
 
-* `PLS` (Partial Least Squares regression) method added
+* `PLS` (Partial Least Squares regression) method added.
 
-* `read` and the more specific `read_remote` method now handle any url pointing
+* `read` method now handle any url pointing
   to a spectrochempy readeable file. An url to a compressed (zip) files are also accepted.
 
-  example:
+  Example:
 
     .. code-block:: ipython
 
@@ -33,6 +33,13 @@ New features
       # lst contains 7 NDDatasets,, display the last
       lst[-1].plot()
 
+* Download from urls can also be done using the `download` method.
+  However `read` offers more options such as merging.
+
+* Automatically download the github repository ``spectrochempy_data`` which contains the files
+  for the examples and tests. The files are downloaded in the directory scp.preferences.datadir.
+  When the program is run for the first time, the availability of the files may take several
+  minutes, depending on the quality of the internet connection.
 
 .. section
 
@@ -48,6 +55,7 @@ Breaking changes
 ~~~~~~~~~~~~~~~~
 .. Add here new breaking changes (do not delete this comment)
 
+* The `read_remote` method has been removed. Use `read` instead.
 
 .. section